AtCoder

+++ include/templates/Graph/Kruskal.hpp
@@ -4,15 +4,15 @@
 
 struct edge {
     int a, b;
-    int c;
+    i64 c;
     friend bool operator<(const edge &lhs, const edge &rhs) {
         return lhs.c < rhs.c;
     }
 };
-int Kruskal(int n, std::vector<edge> edges) {
+i64 Kruskal(int n, std::vector<edge> edges) {
     std::sort(edges.begin(), edges.end());
     DSU dsu(n);
-    int ans = 0;
+    i64 ans = 0;
     for (auto &[a, b, c] : edges) {
         if (dsu.merge(a, b)) {
             ans += c;
